move functions to featurize.py
