+++ src/matbench_genmetrics/utils/match.py
@@ -2,7 +2,6 @@ import logging
 import warnings
 
 import numpy as np
-from matminer.featurizers.composition.composite import ElementProperty
 from matminer.featurizers.site.fingerprint import CrystalNNFingerprint
 from pymatgen.analysis.bond_valence import BVAnalyzer
 from pymatgen.analysis.structure_matcher import StructureMatcher
@@ -70,69 +69,8 @@ def structure_pairwise_match_matrix(
     return match_matrix
 
 
-CompFP = ElementProperty.from_preset("magpie")
-
-
-def cdvae_cov_comp_fingerprints(structures, verbose=False):
-    my_tqdm = get_tqdm(verbose)
-    return [CompFP.featurize(s.composition) for s in my_tqdm(structures)]
-
-
 CrystalNNFP = CrystalNNFingerprint.from_preset("ops")
 bva = BVAnalyzer()
-
-
-def cdvae_cov_struct_fingerprints(structures, verbose=False):
-    """Use SiteStatsFingerprint if OK with NaN rows upon partial site failures."""
-    my_tqdm = get_tqdm(verbose)
-    struct_fps = []
-    num_sites = []
-    num_failed_sites = []
-    failed_structures = []
-    for s in my_tqdm(structures):
-        site_fps = []
-        exceptions = []
-        ns = len(s)
-        num_sites.append(ns)
-        for i in range(ns):
-            try:
-                site_fp = CrystalNNFP.featurize(s, i)
-                site_fps.append(site_fp)
-            except Exception as e:
-                exceptions.append(f"site {i}: {str(e)}")
-
-        num_failed_sites.append(len(exceptions))
-        if exceptions:
-            exception_strs = "\n".join(exceptions)
-            logger.warning(
-                f"{len(exception_strs)} exceptions encountered for structure {i}:\n{s}\n The exceptions are:\n{exception_strs}"  # noqa: E501
-            )
-        if site_fps:
-            struct_fp = np.array(site_fps).mean(axis=0)
-        else:
-            failed_structures.append(s)
-            # NaN vector https://stackoverflow.com/a/1704853/13697228
-            num_features = len(CrystalNNFP.feature_labels())
-            struct_fp = np.empty(num_features)
-            struct_fp[:] = np.nan
-        struct_fps.append(struct_fp)
-    if num_failed_sites:
-        fail_rate = np.array(num_failed_sites) / np.array(num_sites)
-        avg_fail_rate = np.mean(fail_rate[fail_rate > 0])
-        logger.warning(
-            f"{len(num_failed_sites)} structures partially failed to featurize, with on average {avg_fail_rate:.2f} site failure rate per failed structure, and where failed sites were ignored during averaging."  # noqa: E501
-        )
-    if failed_structures:
-        failed_structure_strs = [str(fs) for fs in failed_structures]
-        if len(failed_structure_strs) < 10:
-            failed_structure_str = "\n".join(failed_structure_strs)
-        else:
-            failed_structure_str = "\n".join(failed_structure_strs[:10])
-        logger.warning(
-            f"{len(failed_structures)} structures totally failed to featurize. These were replaced with NaN values. Up to the first 10 structures are displayed here: {failed_structure_str}"  # noqa: E501
-        )
-
-    return struct_fps
 
 
 def cdvae_cov_match_matrix(
